Use sub components on workflow overview page
